Reorganize generators
If we can develop W≠1 Villain-formulation algorithms

    https://github.com/evanberkowitz/supervillain/issues/113
    https://github.com/evanberkowitz/supervillain/issues/112

it will behoove us to keep the generators a bit more organized.  Here I
have separated the generators into Villain and Worldline, which
parallels how the actions themselves are organized.
