+++ supervillain/generator/__init__.py
@@ -4,4 +4,3 @@ from .example import DoNothing
 import supervillain.generator.villain
 import supervillain.generator.worldline
 import supervillain.generator.combining
-import supervillain.generator.worm
